backend-files/requirements.txt
# Interceptor Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
tqdm==4.66.1

# ML Dependencies (optional for full inference)
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# For Hugging Face model downloads
huggingface-hub==0.19.4
